+++ lib/check.js
@@ -9,48 +9,9 @@ var Path = require('path');
 var Wreck = require('wreck');
 var PathIsAbsolute = require('path-is-absolute');
 var SanitizePackage = require('./utils/sanitizePackage');
-var Lockfile = require('@yarnpkg/lockfile');
+var Yarn = require('./utils/yarn');
 
 var internals = {};
-
-internals.transformYarn = function transformYarn(result) {
-
-  internals.processPackages = function processPackages(object) {
-
-    Object.keys(object).forEach(function (packageAtVersion) {
-      // Nested dependencies must be { version: "x" } instead of "x"
-      if (typeof object[packageAtVersion].dependencies === 'object') {
-        internals.deepenDependencies(object[packageAtVersion].dependencies);
-      }
-
-      // Strip version from 'package@^6.1.2'
-      // Handle package scopes like '@types/nsp@1.2.0'
-      var versionAt = packageAtVersion.indexOf('@', 1);
-      if (versionAt >= 1) {
-        var packageName = packageAtVersion.substr(0, versionAt);
-        if (typeof object[packageName] === 'undefined') {
-          object[packageName] = object[packageAtVersion];
-          delete object[packageAtVersion];
-        }
-      }
-    });
-    return object;
-  };
-
-  internals.deepenDependencies = function deepenDependencies(deps) {
-
-    Object.keys(deps).forEach(function (dep) {
-
-      if (typeof deps[dep] === 'string') {
-        deps[dep] = { version: deps[dep] };
-      }
-    });
-  };
-
-  internals.processPackages(result.object);
-  internals.deepenDependencies(result.object);
-  return { dependencies: result.object };
-};
 
 var Conf = require('rc')('nsp', {
   api: {
@@ -175,7 +136,7 @@ module.exports = function (options, callback) {
   if (!options.shrinkwrap) {
     if (options.yarnlock) {
       var yarnlock = Fs.readFileSync(require.resolve(options.yarnlock), 'utf8');
-      options.shrinkwrap = internals.transformYarn(Lockfile.parse(yarnlock));
+      options.shrinkwrap = Yarn.parse(yarnlock);
       shrinkwrap = internals.stringifyShrinkwrap(options.shrinkwrap);
     }
   }
